[TESTS] Refactoring des tests
